requirements.txt
Flask
Flask-CORS
Flask-Bcrypt
Flask-SQLAlchemy
python-dotenv